Fixed the Wolf Speed 'bug' & Level 5 Bug
Wolf Bug:
The speed of the wolves was to fast and were diffined in the XML files.
Had to change them all to speed of 2. --> Speed = 3 to Speed = 2.
Level 5 Bug:
The collision was not set becuase there was no default collision to be
set. (Only when there were enemies there was a method selected) -->
Added a default method: Direction Radar.